pyproject.toml
[project]
name = "enaex"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "azure-servicebus>=7.14.3",
    "azure-storage-blob>=12.27.1",
    "matplotlib>=3.10.7",
    "numpy>=2.3.5",
    "onnxruntime>=1.23.2",
    "opencv-python>=4.11.0.86",
    "python-dotenv>=1.2.1",
    "scikit-image>=0.25.2",
    "shapely>=2.1.2",
]
